refactor: refactor slow addition

+++ tfg/src/main/java/com/example/diagrams/RecursivePowerDiagram.java
@@ -40,7 +40,7 @@ public class RecursivePowerDiagram implements IDiagramActions{
             Map<String,String> returnVal = new HashMap<>();
             returnVal.put("ogSol",String.valueOf(Algorithms.recursiveExponentOption1(a,b,baseCaseValue)));
             returnVal.put("partSol",String.valueOf(Algorithms.recursiveExponentOption1(a,b-1,baseCaseValue)));
-            returnVal.put("reducedOperation",String.format("%d, %d",(int)a,(int)b-1));
+            returnVal.put("reducedOperation",String.format("%.0f, %.0f",a,b-1));
             returnVal.put("currentReductionSolutions",String.valueOf(0));
             return returnVal;
         });
@@ -49,12 +49,12 @@ public class RecursivePowerDiagram implements IDiagramActions{
             returnVal.put("ogSol",String.valueOf(Algorithms.recursiveExponentOption1(a,b,baseCaseValue)));
             if(b%2==0){
                 returnVal.put("partSol",String.valueOf(Algorithms.recursiveExponentOption1(a,b/2,baseCaseValue)));
-                returnVal.put("reducedOperation",String.format("%d,%d",(int)a,(int)b/2));
+                returnVal.put("reducedOperation",String.format("%.0f,%.0f",a,b/2));
                 returnVal.put("currentReductionSolutions",String.valueOf(1));
             }
             else{
                 returnVal.put("partSol",String.valueOf(Algorithms.recursiveExponentOption1(a,(b-1)/2,baseCaseValue)));
-                returnVal.put("reducedOperation",String.format("%d,%d",(int)a,(int)(b-1)/2));
+                returnVal.put("reducedOperation",String.format("%.0f,%.0f",a,(b-1)/2));
                 returnVal.put("currentReductionSolutions",String.valueOf(2));
             }
             return returnVal;
